Adjust name and some terms
There are more with this series coming "Mastering sequential logic in aa/cja: ______" so I suggest here editing the title to include the subtitle "Foundations". And I adjusted a few terms in description and first paragraph to capture more of the varied language used by those in AA context or in CJA context.

I'm the author of the video and just want to suggest these edits to be helpful for the future and any search indexing.